requirements.txt
ollama==0.1.7
langchain==0.1.0
langchain-community==0.0.13
chromadb==0.4.22
psycopg2-binary==2.9.9
pydantic==2.5.3
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
chromadb==0.4.22
sentence-transformers==2.2.2